Resolve all five D3D12 CBV forms for the Fade control tracer; accept write-only mapped buffers
The tracer only recognized plain push_descriptors, but modern engines
(including UE's D3D12 RHI) use root signature 1.1/1.2, which ReShade
classifies as push_descriptors_with_ranges(_and_flags) -- so no real
root-CBV predicate could ever resolve. Unify resolution across all
three push-descriptor forms, and add a diagnostic push_constants
classification for registers backed by root constants rather than a CBV.

Also accept map_access::write_only mapped buffers: D3D12 never reports
read_only/write_discard, so rejecting write_only made every persistent
upload-heap CBV report as not_mapped. Fix mapped-size resolution
(D3D12 always reports UINT64_MAX, not 0) via get_resource_desc.

Investigation-only; Production and the matcher are untouched.

+++ tests/test_fade_control_state.cpp
@@ -65,7 +65,7 @@ FadeControlValueSample Available(float value) {
   return FadeControlValueSample{true, bits, 0};
 }
 
-FadeControlValueSample Unavailable(std::uint8_t reason) {
+FadeControlValueSample Unavailable(std::uint16_t reason) {
   return FadeControlValueSample{false, 0, reason};
 }
